packet: add payload getter
It is not currently possible to access the payload directly from the packet
while it is readily available.
Currently, the alternative is to `packet_rs::slow::parse`, then pop all the
layers to then be able to access the payload with: `packet.to_vec()`.

This change allow to access the packet payload directly from a `Packet` or
`PacketSlice`, making this both work for slow or fast parsing.

Tests are also added to validate the functionality.